(PC-27162) refactor(TrackEmail): Extract content in deprecated component
